Implement nrpe check for failed periodic backups.
This is a fairly large commit that touches several files. The main goal
is to add the ability for nrpe to report to nagios that a periodic
backup has failed to complete. This is accomplished as such: the backup
cron script will drop a periodic_backup.error file in /var/run/ that
includes the error reason. The nrpe check then checks for the existence
of this file, and if it does, set alert to WARNING and include content
of error file. Upon a successful backup, this file will be cleaned up
(including if success from do-backup action).

This commit also includes a few other fixes:

- Add remote_password and private_ssh_key to config.changed for
validation
- Remove check for failed backup flag. No longer blocking on failed
backup
- Add init/update/removal of nrpe checks
- Change flag name from duplicity.failed_periodic_backup to
failed_backup
- Add PidFile functionality to backup script to ensure only 1 runs at
a time. Multiple runs got messy...