layer.yaml: install fabric with layer basic instead of layer venv
The duplicity charm depends on package `fabric`.

1) However, deb package `python3-fabric` is only available on ubuntu 20.04+, so we can not install it via apt layer.
2) Also, the fabric python package needs compiling, so we can not add it in wheelhouse.txt either.

3) The previous implementation was using layer:venv to pip install fabric in venv.
However, layer:venv will never upgrade package according to this line:

    https://github.com/omnivector-solutions/layer-venv/blob/02f6f49949aa5116dc123ad6b1134fd51985b430/lib/charms/layer/venv.py#L41

which is reported in bug:

LP: #1912707

Instead of creating PR on layer-venv and wait for it to merge, we actually do not need it.
layer:basic also has the ability to install python package requires compiling and in venv.

This patch remove layer:venv and install fabric via layer:basic directly.
A missing libssl-dev deb package is also added to fix compiling error.

Signed-off-by: Joe Guo <joe.guo@canonical.com>